1.3.16
Fixed a lot of issues and fixed multiqueue (there's also an update to that file). I also added resampling/motion blur to QM, which is an advanced option, but can be set to always be asked in the options. Another option added was encoding speed, which is set to ultrafast by default. It doesn't change much but could be fun to test with as I haven't tested extensively. I also added a small "ding" sound when the video finishes rendering.